feat: add new function to validate guess and bad request situations
