+++ src/script.js
@@ -1,6 +1,8 @@
 import constants from './constants.js';
+
 import getRightNumber from './services/api.js';
-import getGuess from './functions/guess/index.js';
+
+import guessFunctions from "./functions/guess/index.js";
 import displayFunctions from './functions/display/index.js';
 
 var rightNumber;
@@ -9,6 +11,7 @@ const playAgainButton = document.querySelector(`.${constants.classes.PLAY_AGAIN_
 
 async function initialize()  {
   const resultsContent = document.querySelector(`.${constants.classes.RESULTS_CONTENT}`);
+  const searchInput = document.querySelector(`.${constants.classes.SEARCH_INPUT}`);
 
   displayFunctions.generateDisplay();
 
@@ -16,20 +19,14 @@ async function initialize()  {
   resultsContent.classList.remove( constants.classes.SUCCESS, constants.classes.ERROR);
   playAgainButton.classList.add(constants.classes.HIDDEN);
 
-  let searchInput = document.querySelector(`.${constants.classes.SEARCH_INPUT}`);
-
   sendButton.disabled = false;
   searchInput.disabled = false;
 
   rightNumber = await getRightNumber();
 
   if(!rightNumber.value) {
+    guessFunctions.validateGuess("ERRO", constants.guessStatus.ERROR);
     displayFunctions.generateDisplay(rightNumber.StatusCode);
-    resultsContent.classList.add(constants.classes.ERROR);
-    playAgainButton.classList.toggle(constants.classes.HIDDEN);
-
-    sendButton.disabled = true;
-    searchInput.disabled = true;
   }
 }
 
@@ -38,4 +35,4 @@ document.addEventListener("DOMContentLoaded", () => {
 });
 
 playAgainButton.addEventListener("click", () => initialize());
-sendButton.addEventListener("click", (event) => getGuess(event, rightNumber.value));
+sendButton.addEventListener("click", (event) => guessFunctions.getGuess(event, rightNumber.value));
